% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_aggregated_matrix_to_array.R
\name{convert_aggregated_matrix_to_array}
\alias{convert_aggregated_matrix_to_array}
\title{Convert aggregated 2-dimensional matrix to 3-dimensional array}
\usage{
convert_aggregated_matrix_to_array(mat, group_names = c("pbo", "act"))
}
\arguments{
\item{mat}{Matrix with 4 columns.
Each row contains the data collected in a stratum; hence the number of rows
equals the number of strata.
The data in the columns are required to be as follows:
\itemize{
\item Column 1 - Number of successes in group1,
\item Column 2 - Number of trials in group1,
\item Column 3 - Number of successes in group2, and
\item Column 4 - Number of trials in group2.
}}

\item{group_names}{Character of length 2, indicating names of treatment
groups, defaults to \code{c("pbo", "act")}.}
}
\value{
A 3-dimensional array of stratified 2x2 tables with binomial data obtained
in two independent groups.
The first column of the table contains the outcome \code{y} (e.g. number of
successes or responses) and the second column contains the number of trials
\code{n} (e.g. the number of patients). Each row in the table represents a
sample (e.g. a treatment group).
}
\description{
An auxiliary function to preprocess binomial data collected in 2 groups
across different strata. The input data are \emph{aggregated} at the
stratum-level and are given in a 2-dimensional matrix. The function returns
a 3-dimensional array that is the required input format for other package
functions that estimate common risk differences.
}
\examples{
matrix(
  data = c(3,  7,  1,  4,
           3,  8,  8,  12,
           2,  3,  2,  4,
           2,  5,  2,  6),
  byrow = TRUE,
  ncol = 4
) |>
  convert_aggregated_matrix_to_array()

}
