% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipd_to_array.R
\name{convert_ipd_to_array}
\alias{convert_ipd_to_array}
\title{Convert individual patient-level data to 3-dimensional array}
\usage{
convert_ipd_to_array(
  data,
  stratification,
  treatment,
  endpoint,
  group_names = c("pbo", "act")
)
}
\arguments{
\item{data}{Data frame containing a stratification variable (factor),
a treatment indicator variable (binary) and an endpoint variable (binary).}

\item{group_names}{Character of length 2, indicating names of treatment
groups, defaults to \code{c("pbo", "act")}.}
}
\value{
A 3-dimensional array of stratified 2x2 tables with binomial data obtained
in two independent groups.
The first column of the table contains the outcome \code{y} (e.g. number of
successes or responses) and the second column contains the number of trials
\code{n} (e.g. the number of patients). Each row in the table represents a
sample (e.g. a treatment group).
}
\description{
An auxiliary function to preprocess binomial data collected in 2 groups
across different strata. The input data are \emph{individual patient-level
data} collected across multiple strata, represented in a data frame.
The function returns a 3-dimensional array that is the required
input format for other package functions that estimate common risk
differences.
}
\examples{
data(trialdat)
dat <- trialdat
head(dat)
arr <- convert_ipd_to_array(
  data = dat,
  stratification = c("x1", "x2"),
  treatment = "z",
  endpoint = "y"
)
arr

}
