% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_matrix_to_array.R
\name{convert_matrix_to_array}
\alias{convert_matrix_to_array}
\title{Convert matrix to array}
\usage{
convert_matrix_to_array(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
An array
}
\description{
Convert matrix to array
}
\examples{
matrix(
  data = c(3,  7,  1,  4,
           3,  8,  8,  12,
           2,  3,  2,  4,
           2,  5,  2,  6),
  byrow = TRUE,
  ncol = 4
) |>
  convert_matrix_to_array()

}
\references{
\insertRef{Freedman2008}{sriskdiff}
}
