% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criskdiff-package.R
\docType{package}
\name{criskdiff-package}
\alias{criskdiff}
\alias{criskdiff-package}
\title{criskdiff: Estimation and Inference for the Common Risk Difference in Stratified 2x2 Tables}
\description{
Estimation and Inference for the Common Risk Difference in Stratified 2x2 Tables.
}
\author{
\strong{Maintainer}: Christian Stock \email{christian.stock@boehringer-ingelheim.com} (\href{https://orcid.org/0000-0002-3493-3234}{ORCID})

}
\keyword{internal}
