% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criskdiff_mh_klingenberg.R
\name{criskdiff_mh_klingenberg}
\alias{criskdiff_mh_klingenberg}
\title{Common risk difference with variance estimation using the Klingenberg method}
\usage{
criskdiff_mh_klingenberg(arr, alpha = 0.05)
}
\arguments{
\item{arr}{Array with 3 dimensions containing 2x2 tables, as returned by
\code{\link{convert_matrix_to_array}()}.}

\item{alpha}{Numeric of length 1, level \eqn{\alpha} to compute
\eqn{(1-\alpha)}-confidence intervals.}
}
\value{
A numeric vector with the following elements:
\itemize{
\item \code{est} - the Mantel-Haenszel estimate of the common risk
difference (risk in group 1 (first row) minus risk in group 2
second row)),
\item \code{var} - the variance of \code{est}; \code{NA} for this method,
\item \code{se} - the pseudo-standard error of \code{est},
\item \code{lcl} - the lower \eqn{100(1-\alpha)\%}-confidence interval limit,
\item \code{ucl} - the upper \eqn{100(1-\alpha)\%}-confidence interval limit,
and
\item \code{pval} - the p-value from the test of \eqn{H_0:} risk
difference \eqn{= 0} vs. risk difference \eqn{\ne 0}; \code{NA} for this
method.
}
}
\description{
Computes the Mantel-Haenszel estimate of the common risk difference with
variance estimation according to
\insertCite{Klingenberg2014;textual}{criskdiff}.
}
\details{
The confidence interval may be advantageous when the distribution of risk
differences is skewed. It is not symmetric around the Mantel-Haenszel
estimate of the common risk difference.
}
\examples{
data(myel)
criskdiff_mh_klingenberg(myel)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=convert_matrix_to_array]{convert_matrix_to_array()}} and \code{\link[=criskdiff_mh_sato]{criskdiff_mh_sato()}}.
}
