% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criskdiff_sato.R
\name{criskdiff_sato}
\alias{criskdiff_sato}
\title{Computing risk difference using Sato method}
\usage{
criskdiff_sato(arr)
}
\arguments{
\item{arr}{array}
}
\value{
results
}
\description{
Computing risk difference using Sato method
}
\details{
\insertCite{Sato1989;textual}{criskdiff}
}
\examples{
data(myel)
criskdiff_sato(myel)

}
\references{
\insertAllCited{}
}
