% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criskdiff_score.R
\name{criskdiff_score}
\alias{criskdiff_score}
\title{Score-based common risk difference and interval estimation}
\usage{
criskdiff_score(arr, alpha = 0.05)
}
\arguments{
\item{arr}{Array with 3 dimensions containing 2x2 tables, as returned by
\code{\link{convert_aggregated_matrix_to_array}()} or
\code{\link{convert_ipd_to_array}()}.}

\item{alpha}{Numeric of length 1, level \eqn{\alpha} to compute
\eqn{(1-\alpha)}-confidence intervals.}
}
\value{
A numeric vector with the following elements:
\itemize{
\item \code{est} - the score-based estimate of the common risk
difference (risk in group 1 (first row) minus risk in group 2 (second
row)),
\item \code{var} - the variance of \code{est},
\item \code{se} - the standard error of \code{est},
\item \code{lcl} - the lower \eqn{100(1-\alpha)\%}-confidence interval
limit,
\item \code{ucl} - the upper \eqn{100(1-\alpha)\%}-confidence interval
limit, and
\item \code{pval} - the p-value from the (two-sided) test of
\eqn{H_0:} risk difference \eqn{= 0} vs. risk difference \eqn{\ne 0};
here \code{NA}.
}
}
\description{
Computes a score-based common risk difference estimate and confidence
limits based on stratum risk differences estimated according to
\insertCite{Miettinen1985;textual}{criskdiff} and
\insertCite{Mee1984;textual}{criskdiff}.
}
\examples{
data(myel)
criskdiff_score(myel)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=convert_aggregated_matrix_to_array]{convert_aggregated_matrix_to_array()}},
\code{\link[=criskdiff_mh_sato]{criskdiff_mh_sato()}} and
\code{\link[=criskdiff_mh_klingenberg]{criskdiff_mh_klingenberg()}}.
}
