% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criskdiff_score_mn.R
\name{criskdiff_score_mn}
\alias{criskdiff_score_mn}
\title{Miettinen-Nurminen (Score) Confidence Limits for a Single Risk Difference}
\usage{
criskdiff_score_mn(mat, alpha = 0.05)
}
\arguments{
\item{mat}{A 2x2 matrix}

\item{alpha}{Numeric of length 1, level \eqn{\alpha} to compute
\eqn{(1-\alpha)}-confidence intervals.}
}
\value{
A vector
}
\description{
Computes score-based confidence limits for a single risk difference
according to \insertCite{Miettinen1985;textual}{criskdiff}.
}
\examples{
data(myel)
criskdiff_score_mn(myel)

}
\references{
\insertAllCited{}
}
