% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn_score.R
\name{mn_score}
\alias{mn_score}
\title{Miettinen-Nurminen (Score) Confidence Limits for a Single Risk Difference}
\usage{
mn_score(mat, alpha = 0.05)
}
\arguments{
\item{mat}{A 2x2 matrix}

\item{alpha}{Numeric of length 1, level \eqn{\alpha} to compute
\eqn{(1-\alpha)}-confidence intervals.}
}
\value{
A vector
}
\description{
Computes score-based confidence limits for a single risk difference
according to \insertCite{Miettinen1985;textual}{criskdiff}.
}
\examples{
mat <- matrix(data = c(5, 15, 10, 20),
              nrow = 2,
              byrow = TRUE)
mn_score(mat)

}
\references{
\insertAllCited{}
}
