% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myel}
\alias{myel}
\title{Myeloma dataset}
\format{
A data frame with 53940 rows and 10 variables
}
\source{
\url{https://sites.williams.edu/bklingen/files/2013/06/myel.txt}
}
\usage{
data(myel)
}
\description{
An array containing data from a randomized trial carried out in \eqn{K=21} institutions.
\insertCite{parseRd;Klingenberg2014}{sriskdiff}
}
\details{
\itemize{
\item price.
}
}
\references{
\insertRef{Klingenberg2014}{sriskdiff}
}
\keyword{datasets}
