% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trialdat}
\alias{trialdat}
\title{Simulated trial dataset}
\format{
Object of class "tibble" (90 x 5).
}
\usage{
data(trialdat)
}
\description{
Simulated (hypothetical) individual patient data from a clinical trial with
90 rows (patients) and 5 columns (variables). Columns 1 and 2 (\code{x1} and
\code{x2}) represent binary baseline stratification variables. Column 3
(\code{x3}) represents a continuous baseline variable. Column 4 (\code{z})
is a binary treatment indicator and column 5 (\code{y}) is the binary
response endpoint.

The marginal sample proportions are as follows:\tabular{ccc}{
   Treatment \tab Count \tab \eqn{P(y = 1 | z)} \cr
   \eqn{z = 0} \tab 30 \tab 0.23 \cr
   \eqn{z = 1} \tab 60 \tab 0.55 \cr
}
}
\keyword{datasets}
