% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myel}
\alias{myel}
\title{Myeloma dataset}
\format{
A data frame with 53940 rows and 10 variables
}
\source{
\url{https://sites.williams.edu/bklingen/files/2013/06/myel.txt}
}
\usage{
data(myel)
}
\description{
An array containing data from a randomized controlled trial used in
\insertCite{Klingenberg2014;textual}{criskdiff}.
}
\details{
The trial compared two chemotherapy treatments with respect to survival
(alive or dead by the end of the study; i.e. a binary endpoint) in patients
with multiple myeloma.
It was was carried out in \eqn{K=21} institutions (strata), with an average
of 7.4 patients per institution.

\describe{
\item{group}{First item}
\item{outcome}{Second item}
\item{stratum}{Second item}
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
